feat(envelope): move canonical types to src/envelope.py, add app/core compat re-export
